+++ SecretColors/data/constants.py
@@ -10,9 +10,11 @@ PALETTE_IBM = "ibm"
 PALETTE_MATERIAL = "material"
 PALETTE_BREWER = "brewer"
 PALETTE_CLARITY = "clarity"
+PALETTE_TABLEAU = "tableau"
 
 # Order of following palette will be used to extract the additional color
-ALL_PALETTES = [PALETTE_IBM, PALETTE_MATERIAL, PALETTE_BREWER, PALETTE_CLARITY]
+ALL_PALETTES = [PALETTE_IBM, PALETTE_MATERIAL, PALETTE_BREWER,
+                PALETTE_CLARITY, PALETTE_TABLEAU]
 
 MODE_HEX = "hex"
 MODE_RGB = "rgb"
@@ -25,4 +27,14 @@ MODE_HEX_A = "hexa"  # With Transparency
 ALL_COLOR_MODES = [MODE_HEX, MODE_AHEX, MODE_RGB, MODE_RGBA, MODE_HSL,
                    MODE_HSLA, MODE_HEX_A]
 
-ALL_COLORS = []
+SYNONYM = {
+    "grey": "gray",
+    "r": "red",
+    "b": "blue",
+    "g": "green",
+    "c": "cyan",
+    "m": "magenta",
+    "y": "yellow",
+    "k": "black",
+    "w": "white"
+}