+++ SecretColors/data/names/__init__.py
@@ -1,0 +1,8 @@
+#  Copyright (c) SecretBiology  2020.
+#
+#  Library Name: SecretColors
+#  Author: Rohit Suratekar
+#  Website: https://github.com/secretBiology/SecretColors
+#
+#
+