+++ SecretColors/data/palettes/__init__.py
@@ -4,10 +4,13 @@
 #  Author: Rohit Suratekar
 #  Website: https://github.com/secretBiology/SecretColors
 #
-#
+#  Most of these palettes are derived from various design systems. Few
+#  examples of such design systems can be found on following URL
+#  https://designsystemsrepo.com/design-systems
 
 from SecretColors.data.palettes.parent import ParentPalette
 from SecretColors.data.palettes.ibm import IBMPalette
 from SecretColors.data.palettes.material import MaterialPalette
 from SecretColors.data.palettes.clarity import ClarityPalette
 from SecretColors.data.palettes.brewer import ColorBrewer
+from SecretColors.data.palettes.tableau import TableauPalette