+++ SecretColors/data/palettes/todo.py
@@ -1,0 +1,16 @@
+#  Copyright (c) SecretBiology  2020.
+#
+#  Library Name: SecretColors
+#  Author: Rohit Suratekar
+#  Website: https://github.com/secretBiology/SecretColors
+#
+#  All TODO
+#
+#  Latitude Color Palette
+#  https://www.flexport.com/design/design-language/color-system
+#
+#  Polaris
+#  https://polaris.shopify.com/design/colors#navigation
+#
+#  Adobe Spectrum
+#  https://spectrum.adobe.com/page/color/